Clean up testing framework.
Define libraries libcrypto and libssl even when testing. This way there's no
need for #-:cl+ssl-foreign-libs-already-loaded all over the place.

The package cl+ssl needs to be defined before the rest of the system is loaded
so that the testing framework can define the libraries cl+ssl::libcrypto and
cl+ssl::libssl before loading the system.